added more features
The timer which the user selects is now displayed and working on the session page. Object detection has also been implemented on the camera frames. When a user displays specific objects, a countdown will initiate, if 5 seconds passes and the object is still detectede, the failed session page is displayed. If user reaches end of session without getting distracted, the session success page is shown. all js files have been combined into a single file